Rewrite the Platform class.
It can now be of any size on the 40x30 grid. There is test code in
Game1.cs, although the collision detection has been commented out for now.

There were also a few other changes in Game1.cs as well, such as
a dictionary being used to store any loaded graphic assets, and fixing
the resolution to 800x600.