Some work on the Launcher class (part 2).
The launcher works quite well now. It can hold and launch a ball
based on the direction that it is facing.

For testing purposes, there is a restart button now, which is 'r'.
The ball also has friction, although it really needs to be improved.
And as always, there's other things that I always forget to mention.