Make an object's height and width independent of the texture.
I'm not sure that hardcoding the dimensions of objects is a good idea,
but I didn't want the dimensions of an object to depend on the texture
either. Decoupling the two is for the best, I think.

Also, Git yelled at me because the line endings are different or
something, so I had to fix that first...