Fix platform weirdness.
This commit allows the physics engine to work with all types of
platforms, not just those defined in the original Brian_Collision
branch. This allows the data in the example XML file to be interpreted
properly. The following changes were made:
- The 20x20 texture is now scaled to fit the size of the platform.
- Code has been added to compensate for the fact that the physics
engine treats (width / 2, height / 2) as the position of the platform.
- Due to this, the code that creates the walls was rewritten.