Level Editor
Added 3 new classes to make the level editor work, GameEditor Screen,
ToolboxScreen, and IlevelObject. Things that still need to be done:
Limit the Platforms that can be added, add more platforms to the
Toolbox, or populate depending on level, level saving. The game now
loads directly into the game editor, right click to launch tolbox and
hit enter to go into the game.