Updated Level Select Screen
Added all the levels created to the level select screen. Now you can
play all the levels. The status screen still needs work to show the
correct data. It only really displays the correct name of the level.
Overall the code may need to be cleaned up and comments need to be
added.