Implement a simple UserProfile class.
This class is composed of a dictionary of LevelStatus objects, representing the user's
progress through the game, and a UserPrefs object. It can be [de]serialized with
.NET's XmlSerializer class.

Because the Dictionary provided by the .NET Framework is not XML-serializable,
a serializable subclass was sourced from the Internet. Attribution is included
in a comment at the header of the SerializableDictionary.cs file.