Updated Level Select and Level Status Screens
Level Select now distinguishes between completed and not completed
levels, show which are unlocked based on the previous level. [Note the
first level at the moment needs to be completed, I am going to have it
be unlocked even though not completed. Level Status Screen display time
in minutes and seconds now.