Level select screen displays levels properly and options screen updated
The Level select screen displays locked levels in gray, unlocked levels
in white.
The options screen now has a restart from level editor option that
works. It does not display restart level in editor mode, but does
display it in simulator mode. Also a controls option is available that
displays the controls to the player in the pause menu. It is not very
clear though maybe?